requirements.txt
ipython
numpy
matplotlib==3.5.1
scipy
scikit-image==0.19.3
pandas
pytorch-lightning
einops
omegaconf
protobuf==3.20
tornado
